# [D] SIMD concurrent version