Now it is produced the right answer.
Based on C# code