@jinyus 2023-10-18
Main branch:
```
17.20 ms 
172.17 ms  
1.40 s
```

This PR:
```
D Concurrent:                                                                                                                                                                                                                                  
                                                                                                                                                                                                                                               
        Benchmark 1: ./related_concurrent                                                                                                                                                                                                      
        Processing time (w/o IO): 12.768ms                                                                                                                                                                                                     
        Processing time (w/o IO): 12.231ms                                                                                                                                                                                                     
        Processing time (w/o IO): 12.738ms                                                                                                                                                                                                     
        Processing time (w/o IO): 11.9ms                                                                                                                                                                                                       
        Processing time (w/o IO): 12.854ms
        Processing time (w/o IO): 13.58ms
        Processing time (w/o IO): 11.832ms
        Processing time (w/o IO): 12.737ms
        Processing time (w/o IO): 13.832ms
        Processing time (w/o IO): 12.716ms
        Processing time (w/o IO): 12.202ms
        Processing time (w/o IO): 12.421ms
        Processing time (w/o IO): 13.745ms
          Time (mean ± σ):      91.3 ms ±   1.3 ms    [User: 105.6 ms, System: 9.4 ms]
          Range (min … max):    89.6 ms …  93.7 ms    10 runs

D Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 120.52ms
        Processing time (w/o IO): 120.359ms
        Processing time (w/o IO): 120.118ms
          Time (mean ± σ):     418.5 ms ±   1.6 ms    [User: 695.2 ms, System: 35.7 ms]
          Range (min … max):   417.4 ms … 419.6 ms    2 runs

D Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 953.359ms
        Processing time (w/o IO): 953.074ms
        Processing time (w/o IO): 959.911ms
          Time (mean ± σ):      1.858 s ±  0.003 s    [User: 4.438 s, System: 0.093 s]
          Range (min … max):    1.856 s …  1.861 s    2 runs

```